moved staticmethod decorated functions to core, adapted tests as well
