+++ goodman_pipeline/core/__init__.py
@@ -18,7 +18,10 @@ from .core import (astroscrappy_lacosmic,
                    classify_spectroscopic_data,
                    combine_data,
                    convert_time,
+                   create_master_bias,
+                   create_master_flats,
                    dcr_cosmicray_rejection,
+                   define_trim_section,
                    extraction,
                    extract_fractional_pixel,
                    extract_optimal,
@@ -26,12 +29,15 @@ from .core import (astroscrappy_lacosmic,
                    fractional_sum,
                    get_best_flat,
                    get_central_wavelength,
+                   get_overscan_region,
                    get_slit_trim_section,
                    get_twilight_time,
                    identify_targets,
                    image_overscan,
                    image_trim,
                    interpolate,
+                   is_file_saturated,
+                   name_master_flats,
                    normalize_master_flat,
                    ra_dec_to_deg,
                    read_fits,
@@ -41,4 +47,5 @@ from .core import (astroscrappy_lacosmic,
                    setup_logging,
                    trace,
                    trace_targets,
+                   validate_ccd_region,
                    write_fits)
